+++ src/index.js
@@ -1,21 +1,22 @@
 import postcss from 'postcss';
 import getRulesMatcher from './rulesMatcher';
-import getResetRules from './resetRules';
-import createRuleInjector from './rulesInjectorFabric';
-
+import getReset from './resetRules';
+import createResetRule from './createResetRule';
 
 export default postcss.plugin('postcss-autoreset', (opts = {})=> {
   opts.rulesMatcher = opts.rulesMatcher || 'bem';
   opts.reset = opts.reset || 'initial';
   const rulesMatcher = getRulesMatcher(opts.rulesMatcher);
-  const resetRules = getResetRules(opts.reset);
-  const injectRule = createRuleInjector(resetRules);
+  const reset = getReset(opts.reset);
   return (css) => {
+    const matchedSelectors = [];
     css.walkRules((rule)=> {
-      if ( !rulesMatcher(rule) ) {
-        return;
+      if (rulesMatcher(rule)) {
+        matchedSelectors.push(rule.selector);
       }
-      injectRule(rule);
     });
+    css.prepend(
+      createResetRule(matchedSelectors, reset)
+    );
   };
 });