+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Put all resets in one rule
+- Add virtual source for generated code
 - Change test suite from tape to ava
 - Change test suite from ava to mocha(for more helpfull diffs)
 - Update deps